refactor: 패키지명 변경
- 예약 시간 패키지를 'time'에서 'reservationtime'으로 변경한다.
