+++ src/main/java/roomescape/reservationtime/ReservationTimeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.time;
+package roomescape.reservationtime;
 
 import java.time.LocalTime;
 import java.util.List;
